fix: layout settings render
